Refine TaskDetails UI and implement strict input validation

+++ ProjectManagementSystem/Constants/TimeConfig.cs
@@ -3,7 +3,7 @@
     public static class TimeConfig
     {
         public const int WorkingHoursPerDay = 8;
-        public const double MinutesInHour = 60.0;
+        public const int MinutesInHour = 60;
         public const string DateDisplayFormat = "dd MMM yyyy";
         public const string DateTimeDisplayFormat = "dd MMM yyyy HH:mm";
         public const string ActivityLogFormat = "MMM dd, HH:mm";
